refactor graph.py
